fix: start deferred recovery after ready and detach dispatch cancel
Start deferred reviewer recovery only after MarkReady so requeue work
cannot bypass the safety-floor admission gate while still starting.
Always use context.WithoutCancel for claimed-item dispatch so BeginShutdown
cancel between the pre-check and ProcessClaimedQueueItem cannot strand
running claims. Split scheduler safety-floor tests under the 300-line cap.

Generated-By: looper 0.11.0-beta.1 (runner=fixer, agent=grok-build)
